+++ CHANGELOG.md
@@ -76,6 +76,8 @@ and this project adheres to
 - cosmwasm-storage: Removed, use [cw-storage-plus] instead. ([#1936])
 - cosmwasm-std: Remove `IbcReceiveResponse`'s `Default` implementation. Use
   `IbcReceiveResponse::new` instead. ([#1942])
+- cosmwasm-vm: Remove `InstanceOptions::print_debug` flag. Set your own handler
+  using `Instance::set_debug_handler`. ([#1953])
 
 [cw-storage-plus]: https://github.com/CosmWasm/cw-storage-plus
 [#1875]: https://github.com/CosmWasm/cosmwasm/pull/1875
@@ -83,6 +85,7 @@ and this project adheres to
 [#1896]: https://github.com/CosmWasm/cosmwasm/pull/1896
 [#1936]: https://github.com/CosmWasm/cosmwasm/pull/1936
 [#1942]: https://github.com/CosmWasm/cosmwasm/pull/1942
+[#1953]: https://github.com/CosmWasm/cosmwasm/pull/1953
 
 ## [1.5.0] - 2023-10-31